Vinay, Alekhya | MOBN-462,643 | Move the conflict type eunm to model package
