+++ api/src/main/java/org/openmrs/module/appointments/model/AppointmentConflictType.java
@@ -1,4 +1,4 @@
-package org.openmrs.module.appointments.constants;
+package org.openmrs.module.appointments.model;
 
 public enum AppointmentConflictType {
     SERVICE_UNAVAILABLE, PATIENT_DOUBLE_BOOKING
