+++ api/src/test/java/org/openmrs/module/appointments/service/impl/AppointmentsServiceImplTest.java
@@ -65,8 +65,8 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static org.openmrs.module.appointments.constants.AppointmentConflictType.PATIENT_DOUBLE_BOOKING;
-import static org.openmrs.module.appointments.constants.AppointmentConflictType.SERVICE_UNAVAILABLE;
+import static org.openmrs.module.appointments.model.AppointmentConflictType.PATIENT_DOUBLE_BOOKING;
+import static org.openmrs.module.appointments.model.AppointmentConflictType.SERVICE_UNAVAILABLE;
 import static org.openmrs.module.appointments.constants.PrivilegeConstants.RESET_APPOINTMENT_STATUS;
 import static org.openmrs.module.appointments.helper.DateHelper.getDate;
 import static org.powermock.api.mockito.PowerMockito.mockStatic;
